+++ rubbergoddess.py
@@ -8,13 +8,14 @@ from core import acl, help, rubbercog, output, utils
 from core.config import config
 from repository.database import database
 from repository.database import session
-from repository.database.karma import Karma, Karma_emoji
-from repository.database.seeking import Seeking
-from repository.database.review import Review, ReviewRelevance, Subject
-from repository.database.verification import User
-from repository.database.image import Image
-from repository.database.points import Points
 from repository.database.acl import ACL_group, ACL_rule, ACL_rule_user, ACL_rule_group
+from repository.database.image import Image
+from repository.database.karma import Karma, Karma_emoji
+from repository.database.meme import Meme
+from repository.database.points import Points
+from repository.database.review import Review, ReviewRelevance, Subject
+from repository.database.seeking import Seeking
+from repository.database.verification import User
 from repository.review_repo import ReviewRepository
 
 intents = discord.Intents.none()